Change space ship control
- Now speedSliderDirection only (0,1) not (-1,1)
- Use UIPanGestureRecognizer instead of UISwipeGestureRecognizer
- Use self.physicsBody.velocity instead of self.physicsBody applyForce

+++ EpochRemastered/JZ_ControlView.h
@@ -12,6 +12,6 @@
 @interface JZ_ControlView : UIView
 
 @property (nonatomic) CGPoint JoystickTouchVector;
-@property (nonatomic) int speedSliderDirection;
+@property (nonatomic) CGFloat speedSliderDirection;
 
 @end
